Rewrite query validation
