+++ package.json
@@ -6,7 +6,6 @@
   },
   "dependencies": {
     "dot-prop": "^7.2.0",
-    "is-coordinates": "~2.0.0",
     "lodash.isplainobject": "~4.0.6",
     "mongoose-detective": "~2.0.0",
     "run-series": "~1.1.8",
