+++ src/utils.ts
@@ -1,0 +1,3 @@
+export function isDefined<T>(arg: T | undefined): arg is T {
+  return typeof arg !== "undefined";
+}
